+++ .storybook/main.ts
@@ -22,7 +22,6 @@ module.exports = {
       options: {
         configureJSX: true,
         babelOptions: {},
-        sourceLoaderOptions: null,
       },
     },
     {