+++ src/icons/logos/wordmark.ts
@@ -1,5 +1,6 @@
-import { WordmarkPrimary } from './svgs';
+// Secondary Logo and Wordmark are identical
+import { WordmarkPrimary, LogoSecondary } from './svgs';
 import { IconProps, useIcon } from '../../shared-components/icon';
 
 export default (props: IconProps) =>
-  useIcon(WordmarkPrimary, WordmarkPrimary, props);
+  useIcon(WordmarkPrimary, LogoSecondary, props);