+++ src/shared-components/accordion/thumbnails/index.tsx
@@ -1,12 +1,6 @@
 import React from 'react';
 
-import {
-  Container,
-  ImageContainer,
-  ThumbnailImage,
-  MultiplesContainer,
-  MultiplesText,
-} from './style';
+import Style from './style';
 
 export interface ThumbnailsProps {
   /** An array of image src strings that Accordion.Thumbails will use to render */
@@ -24,24 +18,24 @@ export const Thumbnails = ({ photoSrcs }: ThumbnailsProps) => {
     }
 
     const firstThumbnail = (
-      <ImageContainer>
-        <ThumbnailImage alt="" src={photoSrcs[0]} />
-      </ImageContainer>
+      <Style.ImageContainer>
+        <Style.ThumbnailImage alt="" src={photoSrcs[0]} />
+      </Style.ImageContainer>
     );
 
     let secondThumbnail = null;
 
     if (photoSrcs.length === 2) {
       secondThumbnail = (
-        <ImageContainer>
-          <ThumbnailImage alt="" src={photoSrcs[1]} />
-        </ImageContainer>
+        <Style.ImageContainer>
+          <Style.ThumbnailImage alt="" src={photoSrcs[1]} />
+        </Style.ImageContainer>
       );
     } else if (photoSrcs.length > 2) {
       secondThumbnail = (
-        <MultiplesContainer>
-          <MultiplesText>+{photoSrcs.length - 1}</MultiplesText>
-        </MultiplesContainer>
+        <Style.MultiplesContainer>
+          <Style.MultiplesText>+{photoSrcs.length - 1}</Style.MultiplesText>
+        </Style.MultiplesContainer>
       );
     }
 
@@ -53,5 +47,5 @@ export const Thumbnails = ({ photoSrcs }: ThumbnailsProps) => {
     );
   };
 
-  return <Container>{renderThumbnails()}</Container>;
+  return <Style.Container>{renderThumbnails()}</Style.Container>;
 };