+++ src/shared-components/callout/index.tsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useTheme } from '@emotion/react';
 
 import Style from './style';
-import { COLORS_PROP_TYPES, ThemeColors, ThemeType } from '../../constants';
+import type { ThemeColors, ThemeType } from '../../constants';
 import { isDefined } from '../../utils/isDefined';
 
 export interface CalloutProps {
@@ -12,33 +12,29 @@ export interface CalloutProps {
    */
   children: React.ReactNode;
   /**
-   * Color of the text and icon
-   */
-  color?: ThemeColors;
-  /**
    * Icon displayed inside the callout right aligned
    */
   icon?: React.ReactNode;
   /**
    * Custom prop to draw on preset Callout styles
    */
-  type?: 'success';
+  type?: 'error' | 'success';
 }
 
 /**
  * Pulls a specific styling preset based on available theme values and `type`
  */
-const getCalloutStyles = (
-  theme: ThemeType,
-  color?: ThemeColors,
-  type?: CalloutProps['type'],
-) => {
+const getCalloutStyles = (theme: ThemeType, type?: CalloutProps['type']) => {
+  // Default values if no type provided
   let backgroundColor: ThemeColors = theme.COLORS.infoLight;
-  let textColor = color ?? theme.COLORS.primary;
+  let textColor: ThemeColors = theme.COLORS.primary;
 
   if (type === 'success') {
     backgroundColor = theme.COLORS.successLight;
     textColor = theme.COLORS.success;
+  } else if (type === 'error') {
+    backgroundColor = theme.COLORS.errorLight;
+    textColor = theme.COLORS.error;
   }
 
   return {
@@ -52,16 +48,18 @@ const getCalloutStyles = (
  *
  * `Callout` will cover the entirety of the container that holds it. You may optionally wrap it with `Callout.Container` which will set the `max-width` to `327px`.
  *
- * If you use a glyph as callout icon the recommended dimesions are 48x48 pixels.
+ * If you use a glyph as callout icon the recommended dimesions are 48x48 pixels (which is the default for Glyphs)
  */
-export const Callout = ({ children, color, icon, type }: CalloutProps) => {
+export const Callout = ({ children, icon, type }: CalloutProps) => {
   const theme = useTheme();
-  const { backgroundColor, textColor } = getCalloutStyles(theme, color, type);
+  const { backgroundColor, textColor } = getCalloutStyles(theme, type);
 
   return (
     <Style.CalloutContainer backgroundColor={backgroundColor}>
       <Style.Text textColor={textColor}>{children}</Style.Text>
-      {isDefined(icon) && <Style.Icon iconColor={textColor}>{icon}</Style.Icon>}
+      {isDefined(icon) && icon !== false && (
+        <Style.Icon iconColor={textColor}>{icon}</Style.Icon>
+      )}
     </Style.CalloutContainer>
   );
 };
@@ -70,6 +68,6 @@ Callout.Container = Style.ParentContainer;
 
 Callout.propTypes = {
   children: PropTypes.node.isRequired,
-  color: COLORS_PROP_TYPES,
   icon: PropTypes.node,
+  type: PropTypes.oneOf(['error', 'success']),
 };