+++ src/shared-components/dropdown/desktopDropdown.tsx
@@ -6,14 +6,7 @@ import { isDefined } from '../../utils/isDefined';
 import { keyboardKeys } from '../../constants/keyboardKeys';
 import { OffClickWrapper } from '../offClickWrapper';
 import { ChevronIcon } from '../../icons';
-import {
-  DropdownContainer,
-  DropdownFocusContainer,
-  dropdownInputStyle,
-  IconContainer,
-  DropdownOptionsContainer,
-  DropdownOption,
-} from './style';
+import Style from './style';
 
 import { OptionType, OptionValue } from './index';
 
@@ -69,8 +62,8 @@ export const DesktopDropdown = <T extends OptionType>({
         width: 100%;
       `}
     >
-      <DropdownContainer textAlign={textAlign}>
-        <DropdownFocusContainer
+      <Style.DropdownContainer textAlign={textAlign}>
+        <Style.DropdownFocusContainer
           onClick={toggleDropdown}
           onKeyDown={handleKeyDown}
           tabIndex={0}
@@ -79,7 +72,7 @@ export const DesktopDropdown = <T extends OptionType>({
           ref={initialFocus}
         >
           <div
-            css={dropdownInputStyle({
+            css={Style.dropdownInputStyle({
               borderRadius,
               shouldBeFullyRounded: !isOpen,
               textAlign,
@@ -88,12 +81,12 @@ export const DesktopDropdown = <T extends OptionType>({
           >
             {currentOption?.label}
           </div>
-          <IconContainer>
+          <Style.IconContainer>
             <ChevronIcon width={10} height={10} rotate={isOpen ? 90 : 0} />
-          </IconContainer>
-        </DropdownFocusContainer>
+          </Style.IconContainer>
+        </Style.DropdownFocusContainer>
 
-        <DropdownOptionsContainer
+        <Style.DropdownOptionsContainer
           borderRadius={borderRadius}
           isOpen={isOpen}
           optionsContainerMaxHeight={optionsContainerMaxHeight}
@@ -114,7 +107,7 @@ export const DesktopDropdown = <T extends OptionType>({
               : `undefined-${index}`;
 
             return (
-              <DropdownOption
+              <Style.DropdownOption
                 key={id}
                 value={optionValue}
                 id={id}
@@ -130,11 +123,11 @@ export const DesktopDropdown = <T extends OptionType>({
                 {...rest}
               >
                 {label}
-              </DropdownOption>
+              </Style.DropdownOption>
             );
           })}
-        </DropdownOptionsContainer>
-      </DropdownContainer>
+        </Style.DropdownOptionsContainer>
+      </Style.DropdownContainer>
     </OffClickWrapper>
   );
 };