+++ src/shared-components/loadingSpinner/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useTheme } from '@emotion/react';
 
-import { LoadingSpinnerContainer, Overlay, Dot } from './style';
-import { ThemeColors } from '../../constants';
+import Style from './style';
+import type { ThemeColors } from '../../constants';
 
 export interface LoadingSpinnerProps {
   /**
@@ -44,28 +44,28 @@ export const LoadingSpinner = ({
   const colorWithTheme = color ?? theme.COLORS.primary;
 
   return (
-    <LoadingSpinnerContainer bgColor={bgColorWithTheme}>
-      <Overlay>
-        <Dot
+    <Style.LoadingSpinnerContainer bgColor={bgColorWithTheme}>
+      <Style.Overlay>
+        <Style.Dot
           dotColor={colorWithTheme}
           translateX={translateX}
           duration={duration}
           dotSize={size}
         />
-        <Dot
+        <Style.Dot
           dotColor={colorWithTheme}
           translateX={translateX}
           duration={duration}
           dotSize={size}
         />
-        <Dot
+        <Style.Dot
           dotColor={colorWithTheme}
           translateX={translateX}
           duration={duration}
           dotSize={size}
         />
-      </Overlay>
-    </LoadingSpinnerContainer>
+      </Style.Overlay>
+    </Style.LoadingSpinnerContainer>
   );
 };