+++ stories/progressBar/index.stories.tsx
@@ -13,7 +13,7 @@ import {
   Title,
 } from '@storybook/addon-docs/blocks';
 import type { Meta } from '@storybook/react';
-import { ThemeColors } from 'src/constants/themes/types';
+import type { ThemeColors } from 'src/constants/themes/types';
 
 const BarContainer = styled.div`
   position: relative;